Edit ingredient amount for recipe
- added dialog to edit ingredient amount for recipe
- change unit displayed to be consistent with pantry for shopping list and recipe
- changed amount displayed for large numbers

+++ app/src/main/java/com/ala158/magicpantry/arrayAdapter/ShoppingListArrayAdapter.kt
@@ -47,7 +47,7 @@ class ShoppingListArrayAdapter(
 
         val shoppingListItem = shoppingListItemAndIngredient[position].shoppingListItem
         val ingredient = shoppingListItemAndIngredient[position].ingredient
-        amountTextView.text = shoppingListItem.itemAmount.toString()
+        amountTextView.text = shoppingListItem.itemAmount.toBigDecimal().toString()
         unitTextView.text = ingredient.unit
         nameTextView.text = ingredient.name
